Moving files into a more sensible directory structure
Added promiseWrapper information to the README.md

Moving files into a more sensible directory structure

added info abou testing to README.md

Modifying paths in preparation for moving files

moving files and adding new index.js in project root

moving test location around

Correcting sourcing o nodeSSH library

corrections due to restructuring

Updating readme for the new structure
