+++ index.js
@@ -1,0 +1,8 @@
+'use strict';
+const promiseWrapper = require('./lib/promiseWrapper');
+const NodeSSH = require('./lib/NodeSSH');
+
+module.exports = {
+  promiseWrapper: promiseWrapper,
+  NodeSSH: NodeSSH
+};
