+++ lib/promiseWrapper.js
@@ -17,7 +17,7 @@ class SSH {
     this.connected = false;
     this.connection = null;
     this.options = options;
-    this.connection = require('./lib/index.js')(this.options);
+    this.connection = require('./NodeSSH.js')(this.options);
     this.connection.on('stderr', function(err) { console.error('Error on StdErr',err); });
   }
